Removing ununsed code and making it so directory paths are passed to scripts from the train.py
