+++ recognition/README.md
@@ -1,10 +1,32 @@
-# Recognition Tasks
-Various recognition tasks solved in deep learning frameworks.
+# Classification Siamese Network Model for ISIC 2020
 
-Tasks may include:
-* Image Segmentation
-* Object detection
-* Graph node classification
-* Image super resolution
-* Disease classification
-* Generative modelling with StyleGAN and Stable Diffusion
+**Author:** Sebastian Dobson, 47447860
+
+## Project Overview
+
+project stuff with images
+
+### Deep Learning Pipeline
+
+1. Data Pre-Processing
+2. Model Architecture
+3. Training Process
+4. Testing Process
+
+## 1. Data Pre-Processing
+
+blah blah
+
+## 2. Model Architecture
+
+blah blah
+
+## 3. Training Process
+
+blah blah
+
+## 4. Testing Process
+
+blah blah - include results
+
+## Project Structure
