+++ recognition/SiameseNet_47447860/predict.py
@@ -10,21 +10,20 @@ from torch.utils.data import DataLoader
 from torchvision import transforms
 
 from modules import SiameseNetwork
-from dataset_3 import Dataset
+from dataset import Dataset
 
 class PredictData:
-    def __init__(self, test_data, show_images=False):
+    def __init__(self, test_data, path, show_images=False):
         self.test_data = test_data
+        self.path = path
         self.show_images = show_images
 
         # results path where our final plots go
-        self.results_path = r"C:\Users\sebas\project\results"
-        #self.results_path = os.path.expanduser("~/project/results/")
+        self.results_path = os.path.join(self.path, "results")
 
         # The path to the model's checkpoint - where weights are saved
         # The checkpoint is kind of like a list of different checkpoints, hence why we need to index it with 'backbone'
-        self.checkpoint_path = r"C:\Users\sebas\project\outputs\best.pth"
-        #self.checkpoint_path = os.path.expanduser("~/project/outputs/best.pth")
+        self.checkpoint_path = os.path.join(self.path, "outputs", "best.pth")
 
         # Set device to CUDA if a CUDA device is available, else CPU
         self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
